iot-app/mqtt-backend/requirements.txt
pymongo==4.11.1
paho-mqtt==2.1.0
